eCourseAppv4/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cloudinary==1.39.0
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
mysqlclient==2.2.4
pillow==10.2.0
PyMySQL==1.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
